ai_agent/frontend.py
import streamlit as st

st.set_page_config(page_title="LangGraph Agent UI", layout="centered")
st.title("AI Chatbot Agents")
st.write("Create and Interact with the AI Agents!")

model_names =["llama3-70b-8192", "gemma2-9b-it", "llama-3.3-70b-versatile"]

selected_model = st.selectbox("Select Model:", model_names)

allow_web_search=st.checkbox("Allow Web Search")

user_query=st.text_area("Enter your query: ", height=150, placeholder="Ask Anything!")

API_URL="http://127.0.0.1:5000/chat"

if st.button("Ask Agent!"):
    if user_query.strip():
        
        import requests

        payload={
            "model_name": selected_model,
            "messages": [user_query],
            "allow_search": allow_web_search
        }

        response=requests.post(API_URL, json=payload)
        if response.status_code == 200:
            response_data = response.json()
            if "error" in response_data:
                st.error(response_data["error"])
            else:
                st.subheader("Agent Response")
                st.markdown(f"**Final Response:** {response_data}")