requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_openai
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
youtube_transcript_api
langgraph
langchain-tavily
pydantic
